refactor wizards
