+++ src/wizards/install-firmware/wizard.ts
@@ -14,11 +14,7 @@ import {
 	type FirmwareType,
 } from "../../lib/firmware-download";
 import { DriverMode } from "zwave-js";
-import {
-	RFRegion,
-	getLegalPowerlevelMesh,
-	getLegalPowerlevelLR,
-} from "@zwave-js/core/definitions";
+import { applyRepeaterRegion } from "../../lib/regions";
 import { type BytesView } from "@zwave-js/shared";
 
 export type { FirmwareType } from "../../lib/firmware-download";
@@ -113,19 +109,6 @@ async function handleFileSelectStepEntry(
 	}
 }
 
-const CLI_REGION_TO_RF_REGION: Record<string, RFRegion> = {
-	EU: RFRegion.Europe,
-	US: RFRegion.USA,
-	ANZ: RFRegion["Australia/New Zealand"],
-	HK: RFRegion["Hong Kong"],
-	IN: RFRegion.India,
-	IL: RFRegion.Israel,
-	RU: RFRegion.Russia,
-	CN: RFRegion.China,
-	JP: RFRegion.Japan,
-	KR: RFRegion.Korea,
-};
-
 async function handleConfigureStepEntry(
 	context: WizardContext<InstallFirmwareState>,
 ): Promise<void> {
@@ -161,7 +144,6 @@ async function handleConfigureBeforeNavigate(
 		return false;
 	}
 
-	// Show spinner
 	context.setState((prev) => ({
 		...prev,
 		configureStatus: "configuring",
@@ -169,59 +151,17 @@ async function handleConfigureBeforeNavigate(
 	}));
 
 	try {
-		const setSuccess = await context.zwaveBinding.setRegion(selectedRegion);
-		if (!setSuccess) {
+		const result = await applyRepeaterRegion(
+			context.zwaveBinding,
+			selectedRegion,
+		);
+		if (result !== true) {
 			context.setState((prev) => ({
 				...prev,
 				configureStatus: "error",
-				configureError: "Failed to set RF region. Please try again.",
+				configureError: result,
 			}));
 			return false;
-		}
-
-		// Wait for ZWA-2 to reboot
-		const { wait } = await import("alcalzone-shared/async");
-		await wait(1000);
-
-		// Verify region
-		const confirmedRegion = await context.zwaveBinding.getRegion();
-		if (confirmedRegion !== selectedRegion) {
-			context.setState((prev) => ({
-				...prev,
-				configureStatus: "error",
-				configureError: `Region verification failed. Expected "${selectedRegion}" but got "${confirmedRegion ?? "unknown"}".`,
-			}));
-			return false;
-		}
-
-		// Configure power levels if legal limits are known for this region
-		const rfRegion = CLI_REGION_TO_RF_REGION[selectedRegion];
-		if (rfRegion != null) {
-			const meshLimit = getLegalPowerlevelMesh(rfRegion);
-			const lrLimit = getLegalPowerlevelLR(rfRegion);
-
-			if (meshLimit != null || lrLimit != null) {
-				// Get current power levels to preserve the adjust value
-				const current =
-					await context.zwaveBinding.getPowerlevel();
-				if (current) {
-					// Z-Wave JS returns dBm, CLI uses deci-dBm
-					const newMeshMax =
-						meshLimit != null
-							? meshLimit * 10
-							: current.txPowerMax;
-					const newLRMax =
-						lrLimit != null
-							? lrLimit * 10
-							: current.txPowerMaxLR;
-
-					await context.zwaveBinding.setPowerlevel(
-						newMeshMax,
-						current.txPowerAdjust,
-						newLRMax,
-					);
-				}
-			}
 		}
 
 		context.setState((prev) => ({
